// src/api/stetsom/model/template.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */
import type { I18nString } from "./i18nString";
import type { TemplateAttr } from "./templateAttr";

export interface Template {
  id: string;
  category_id: string;
  name: I18nString;
  attributes: TemplateAttr[];
  created_at: string;
  updated_at: string;
}