// src/api/stetsom/model/postApiTemplatesBody.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */
import type { PostApiTemplatesBodyName } from "./postApiTemplatesBodyName";
import type { TemplateAttrInput } from "./templateAttrInput";

export type PostApiTemplatesBody = {
  category_id: string;
  name: PostApiTemplatesBodyName;
  attributes?: TemplateAttrInput[];
};