// src/api/stetsom/model/patchApiTemplatesIdBody.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */
import type { PatchApiTemplatesIdBodyName } from "./patchApiTemplatesIdBodyName";
import type { TemplateAttrInput } from "./templateAttrInput";

export type PatchApiTemplatesIdBody = {
  category_id?: string;
  name?: PatchApiTemplatesIdBodyName;
  attributes?: TemplateAttrInput[];
};